+++ pages/carousel.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Carousel = () => {
+  return <div>캐러셀 영역입니다.</div>;
+};
+
+export default Carousel;
