+++ pages/index.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const App = () => {
+  return <p>메뉴를 선택해주세요</p>;
+};
+
+export default App;
