+++ pages/navigation.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Navigation = () => {
+  return <p>네비게이션 영역입니다.</p>;
+};
+
+export default Navigation;
