+++ src/styles/GlobalStyle.js
@@ -1,6 +1,6 @@
 import { css, Global } from "@emotion/react";
 
-const GlobalStyle = () => (
+export const GlobalStyle = () => (
   <Global
     styles={css`
       html,
@@ -64,5 +64,3 @@ const GlobalStyle = () => (
     `}
   />
 );
-
-export default GlobalStyle;
